new-button test has been finished

+++ routes/task.js
@@ -26,27 +26,39 @@ function newTask(req, res){
 
 function getAllTasks(req, res){
 
-    let Task = mongoose.model("Task", taskSchema, req.query.collectionId) 
+    if(req.query.collectionId){
+        let Task = mongoose.model("Task", taskSchema, req.query.collectionId) 
 
-    Task.find()
-    .then((result) => {
-        res.status(200)
-        res.render('tasks', {title: "Tasks", tasks: result})
-        console.log(result)
-    })
-    .catch((error) => console.log(error))
+        Task.find()
+        .then((result) => {
+            res.status(200)
+            res.render('tasks', {title: "Tasks", tasks: result})
+            console.log(result)
+        })
+        .catch((error) => console.log(error))
+    }else{
+        res.render('404', {title: "404 page not found"})
+        res.status(404)
+    }
 }
 
 function deleteTask(req, res){
 
-    Task.findOneAndDelete(req.body.id)
-    .then((task) => {
-        res.status(200)
-        console.log(task)
-    })
-    .catch((error)=>{
-        console.log(error)
-      })
+    if(req.query.collectionId){
+        let Task = mongoose.model("Task", taskSchema, req.query.collectionId) 
+
+        Task.findOneAndDelete(req.body.id)
+        .then((task) => {
+            res.status(200)
+            console.log(task)
+        })
+        .catch((error)=>{
+            console.log(error)
+        })
+    }else{
+        res.render('404', {title: "404 page not found"})
+        res.status(404)
+    }
 }
 
 module.exports = {
